fixed testid size

+++ stm32f413-drivers/filegen/can_types_new.h
@@ -1152,7 +1152,7 @@ typedef struct {
 } cmr_canMemoratorHeartbeat_t;
 
 typedef struct {
-	uint32_t test_id;
+	uint8_t test_id;
 } cmr_canTestID_t;
 
 typedef struct {
